requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.40.0
openai>=1.0.0
python-dotenv==1.0.1
websockets==13.1
pytest==8.3.4
pytest-asyncio==0.24.0
